+++ material-intro-screen/src/main/java/agency/tango/materialintroscreen/MaterialIntroActivity.java
@@ -27,14 +27,15 @@ import agency.tango.materialintroscreen.animations.wrappers.NextButtonTranslatio
 import agency.tango.materialintroscreen.animations.wrappers.PageIndicatorTranslationWrapper;
 import agency.tango.materialintroscreen.animations.wrappers.SkipButtonTranslationWrapper;
 import agency.tango.materialintroscreen.animations.wrappers.ViewPagerTranslationWrapper;
+import agency.tango.materialintroscreen.listeners.IFinishListener;
 import agency.tango.materialintroscreen.listeners.IPageScrolledListener;
 import agency.tango.materialintroscreen.listeners.IPageSelectedListener;
 import agency.tango.materialintroscreen.listeners.MessageButtonBehaviourOnPageSelected;
-import agency.tango.materialintroscreen.listeners.SwipeStateTouchListener;
 import agency.tango.materialintroscreen.listeners.ViewBehavioursOnPageChangeListener;
 import agency.tango.materialintroscreen.listeners.clickListeners.PermissionNotGrantedClickListener;
 import agency.tango.materialintroscreen.listeners.scrollListeners.ParallaxScrollListener;
 import agency.tango.materialintroscreen.widgets.InkPageIndicator;
+import agency.tango.materialintroscreen.widgets.OverScrollViewPager;
 import agency.tango.materialintroscreen.widgets.SwipeableViewPager;
 
 import static android.view.View.GONE;
@@ -49,6 +50,7 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
     private CoordinatorLayout coordinatorLayout;
     private Button messageButton;
     private LinearLayout navigationView;
+    private OverScrollViewPager overScrollLayout;
 
     private ArgbEvaluator argbEvaluator = new ArgbEvaluator();
 
@@ -75,7 +77,9 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
         }
 
         setContentView(R.layout.activity_material_intro);
-        viewPager = (SwipeableViewPager) findViewById(R.id.view_pager_slides);
+
+        overScrollLayout = (OverScrollViewPager) findViewById(R.id.view_pager_slides);
+        viewPager = overScrollLayout.getOverScrollView();
         pageIndicator = (InkPageIndicator) findViewById(R.id.indicator);
         backButton = (ImageButton) findViewById(R.id.button_back);
         nextButton = (ImageButton) findViewById(R.id.button_next);
@@ -101,7 +105,7 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
         viewPager.post(new Runnable() {
             @Override
             public void run() {
-                if (adapter.slidesCount() == 0) {
+                if (adapter.getCount() == 0) {
                     finish();
                 } else {
                     int currentItem = viewPager.getCurrentItem();
@@ -117,7 +121,7 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
         SlideFragment fragment = adapter.getItem(viewPager.getCurrentItem());
         boolean hasPermissionToGrant = fragment.hasNeededPermissionsToGrant();
         if (!hasPermissionToGrant) {
-            viewPager.setAllowedSwipeDirection(SwipeableViewPager.SwipeDirection.all);
+            viewPager.setSwipingRightAllowed(true);
             nextButtonBehaviour(viewPager.getCurrentItem(), fragment);
             messageButtonBehaviourOnPageSelected.pageSelected(viewPager.getCurrentItem());
         } else {
@@ -132,7 +136,7 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
         if (viewPager.getCurrentItem() == 0) {
             finish();
         } else {
-            viewPager.setCurrentItem(viewPager.getPreviousItem());
+            viewPager.setCurrentItem(viewPager.getPreviousItem(), true);
         }
     }
 
@@ -153,12 +157,18 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
     private void initOnPageChangeListeners() {
         messageButtonBehaviourOnPageSelected = new MessageButtonBehaviourOnPageSelected(messageButton, adapter, messageButtonBehaviours);
 
-        viewPager.registerOnTouchEventListener(new SwipeStateTouchListener(viewPager, adapter));
-
         backButtonTranslationWrapper = new BackButtonTranslationWrapper(backButton);
         pageIndicatorTranslationWrapper = new PageIndicatorTranslationWrapper(pageIndicator);
         viewPagerTranslationWrapper = new ViewPagerTranslationWrapper(viewPager);
         skipButtonTranslationWrapper = new SkipButtonTranslationWrapper(skipButton);
+
+        overScrollLayout.registerFinishListener(new IFinishListener() {
+            @Override
+            public void doOnFinish() {
+                onFinish();
+                finish();
+            }
+        });
 
         viewPager.addOnPageChangeListener(new ViewBehavioursOnPageChangeListener(adapter)
                 .registerViewTranslationWrapper(nextButtonTranslationWrapper)
@@ -174,7 +184,7 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
                             @Override
                             public void run() {
                                 if (adapter.getItem(position).hasNeededPermissionsToGrant() || !adapter.getItem(position).canMoveFurther()) {
-                                    viewPager.setCurrentItem(position);
+                                    viewPager.setCurrentItem(position, true);
                                     pageIndicator.clearJoiningFractions();
                                 }
                             }
@@ -216,7 +226,7 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
                         nextButtonTranslationWrapper.error();
                         showError(fragment.cantMoveFurtherErrorMessage());
                     } else {
-                        viewPager.setCurrentItem(position + 1);
+                        viewPager.setCurrentItem(position + 1, true);
                     }
                 }
             });
@@ -268,14 +278,14 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
         skipButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                for (int position = viewPager.getCurrentItem(); position < adapter.slidesCount(); position++) {
+                for (int position = viewPager.getCurrentItem(); position < adapter.getCount(); position++) {
                     if (!adapter.getItem(position).canMoveFurther()) {
-                        viewPager.setCurrentItem(position);
+                        viewPager.setCurrentItem(position, true);
                         showError(adapter.getItem(position).cantMoveFurtherErrorMessage());
                         return;
                     }
                 }
-                viewPager.setCurrentItem(adapter.getLastItemPosition());
+                viewPager.setCurrentItem(adapter.getLastItemPosition(), true);
             }
         });
     }
@@ -290,7 +300,7 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
         backButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                viewPager.setCurrentItem(viewPager.getPreviousItem());
+                viewPager.setCurrentItem(viewPager.getPreviousItem(), true);
             }
         });
     }
@@ -355,10 +365,11 @@ public abstract class MaterialIntroActivity extends AppCompatActivity {
 
     /**
      * Set if last screen should be able to exit with alpha transition
+     * @param enableAlphaExitTransition should enable alpha exit transition
      */
     @SuppressWarnings("unused")
     public void enableLastSlideAlphaExitTransition(boolean enableAlphaExitTransition) {
-        adapter.addEmptySlide(new LastEmptySlideFragment());
+        viewPager.alphaExitTransitionEnabled(enableAlphaExitTransition);
     }
 
     /**